+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed the issue with matching subdomain wildcard rules, e.g. `sub.example.*`,
+  that were not matched before, this is the second fix for [#106].
+
 [unreleased]: https://github.com/AdguardTeam/SafariConverterLib/compare/v4.0.2...HEAD
 
 ## [v4.0.2]